requirements.txt
# This is like a package.json, will pre-install when a container spins up
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv